Move boolean exercises to own package
