+++ app/src/main/java/org/simonolander/lambda/ui/ExerciseDesignView.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import org.simonolander.lambda.domain.Exercise
-import org.simonolander.lambda.domain.andExercise
+import org.simonolander.lambda.content.exercise.booleans.andExercise
 import org.simonolander.lambda.engine.Expression
 import org.simonolander.lambda.engine.ParserException
 import org.simonolander.lambda.engine.parse
